requirements.txt
fastapi
uvicorn
httpx
pyyaml
python-dotenv
sqlalchemy
sentence-transformers